Also evaluate macro helpers when interpreted (#9)
If @-mixins.el is loaded or compiled when @.el is not compiled, @--walk
is undefined. The Makefile prevents this from happening, but package.el,
via byte-recompile-directory, compiles in lexicographic order according
to directory-files, which sorts @-mixins.el first. So also define these
functions during "eval".

A cl-eval-when of (eval load compile) is not equivalent to progn, but I
do not understand why.